+++ R/registry.R
@@ -43,17 +43,19 @@ block_name <- block_descrs_getter(block_descr_getter("name"))
 block_descr <- block_descrs_getter(block_descr_getter("description"))
 
 new_block_descr <- function(constructor, name, description, classes, input,
-                            output, pkg) {
+                            output, pkg, category) {
 
   stopifnot(
     is.function(constructor), is_string(name), is_string(description),
+    is_string(category),
     is.character(classes), length(classes) >= 1L,
     is_string(input), is_string(output), is_string(pkg)
   )
 
   structure(
-    constructor, name = name, description = description, classes = classes,
-    input = input, output = output, package = pkg, class = "block_descr"
+    constructor, name = name, description = description,
+    classes = classes, input = input, output = output,
+    package = pkg, category = category, class = "block_descr"
   )
 }
 
@@ -64,14 +66,24 @@ block_registry <- new.env()
 #' @param classes Block classes
 #' @param input,output Object types the block consumes and produces
 #' @param package Package where block is defined
+#' @param category Useful to sort blocks by topics. If not specified,
+#' blocks are uncategorized.
 #'
 #' @rdname available_blocks
 #' @export
-register_block <- function(constructor, name, description, classes, input,
-                           output, package = NA_character_) {
+register_block <- function(
+  constructor,
+  name,
+  description,
+  classes,
+  input,
+  output,
+  package = NA_character_,
+  category = "uncategorized"
+) {
 
   descr <- new_block_descr(constructor, name, description, classes, input,
-                           output, package)
+                           output, package, category)
 
   id <- classes[1L]
 
@@ -85,8 +97,16 @@ register_block <- function(constructor, name, description, classes, input,
 #' @param ... Forwarded to `register_block()`
 #' @rdname available_blocks
 #' @export
-register_blocks <- function(constructor, name, description, classes, input,
-                            output, package = NA_character_) {
+register_blocks <- function(
+  constructor,
+  name,
+  description,
+  classes,
+  input,
+  output,
+  package = NA_character_,
+  category = "uncategorized"
+) {
 
   if (length(constructor) == 1L && !is.list(classes)) {
     classes <- list(classes)
@@ -100,7 +120,8 @@ register_blocks <- function(constructor, name, description, classes, input,
     classes = classes,
     input = input,
     output = output,
-    package = package
+    package = package,
+    category = category
   )
 
   invisible(res)
@@ -256,7 +277,25 @@ register_blockr_blocks <- function(pkg) {
       "data.frame",
       "data.frame"
     ),
-    package = pkg
+    package = pkg,
+    category = c(
+      "data",
+      "data",
+      "data",
+      "data",
+      "parser",
+      "parser",
+      "parser",
+      "parser",
+      "transform",
+      "transform",
+      "transform",
+      "transform",
+      "transform",
+      "transform",
+      "transform",
+      "transform"
+    )
   )
 }
 
@@ -272,4 +311,31 @@ construct_block <- function(block, ...) {
   stopifnot(inherits(block, "block_descr"))
 
   block(...)
+}
+
+#' List available blocks as a data.frame
+#'
+#' Provides an alternate way of displaying
+#' the registry information.
+#' This can be useful to create dynamic UI elements
+#' like in \link{add_block_ui}.
+#'
+#' @return A dataframe.
+#'
+#' @export
+get_registry <- function() {
+  res <- lapply(seq_along(available_blocks()), \(i) {
+    blk <- available_blocks()[[i]]
+    attrs <- attributes(blk)
+    data.frame(
+      ctor = names(available_blocks())[[i]],
+      description = attrs[["description"]],
+      category = attrs[["category"]],
+      classes = paste(c(attrs[["classes"]], "block"), collapse = ", "),
+      input = attrs[["input"]],
+      output = attrs[["output"]],
+      package = attrs[["package"]]
+    )
+  })
+  do.call(rbind, res)
 }